fix multiple definiton hell in Numbers.h and Utilities.h
separated utilities definition in .cc and .h. Make numbers
function inlined.
